poc1 v0.1.0 Poc1

Ejemplo básico de pattern matching y envío de mensajes entre procesos

Functions

contesta(pid, pregunta, n)

Responde ante una pregunta ping

##Ejemplo
iex>{_pid, resp, n} = Poc1.contesta(self(), "ping", 10)
iex>resp
"pong"
iex>n
11

escucha()

Función recursiva encargada de mantener el proceso a la escucha

inicia_dialogo()

Lanza en un proceso distinto la función escucha, tras lo cual le envía un mensaje y se pone a la escucha